Fix client and server bugs

+++ Server/src/main/java/com/mwojnar/ChatImpl.java
@@ -7,21 +7,23 @@ import io.grpc.StatusRuntimeException;
 import io.grpc.stub.StreamObserver;
 
 import java.io.IOException;
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.sql.Types;
-import java.util.*;
+import java.sql.*;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.Map;
 import java.util.logging.Logger;
 
 public class ChatImpl extends ChatGrpc.ChatImplBase {
     private final Logger logger;
     private final Connection connection;
     private final Map<Long, StreamObserver<Message>> observers;
+    private final Map<Long, Long> groups;
 
     public ChatImpl(Logger logger, Connection connection) {
         this.logger = logger;
         this.connection = connection;
         this.observers = new HashMap<>();
+        this.groups = new HashMap<>();
     }
 
     @Override
@@ -55,11 +57,12 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
                 .setClientAckId(ackId)
                 .build();
 
-        logger.info("User " + request.getUserId() + " joined successfully.");
+        logger.info("User " + request.getUserId() + " joined group " + request.getGroupId() + " successfully.");
         responseObserver.onNext(response);
+        responseObserver.onCompleted();
     }
 
-    private void createGroupIfNotExists(long groupId) throws SQLException {
+    private synchronized void createGroupIfNotExists(long groupId) throws SQLException {
         var statement = connection.createStatement();
         var resultSet = statement.executeQuery("SELECT * FROM groups WHERE groups.group_id == " + groupId);
 
@@ -69,7 +72,7 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
         }
     }
 
-    private long insertOrUpdateUser(long userId, long groupId, Long ackId) throws SQLException {
+    private synchronized long insertOrUpdateUser(long userId, long groupId, Long ackId) throws SQLException {
         var statement = connection.createStatement();
 
         if (ackId == null) {
@@ -81,7 +84,7 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
         return ackId;
     }
 
-    private long getLastTime(long userId) throws SQLException {
+    private synchronized long getLastTime(long userId) throws SQLException {
         var statement = connection.createStatement();
         var resultSet = statement.executeQuery("SELECT MAX(time) FROM messages WHERE user_id == " + userId);
         return resultSet.getLong(1);
@@ -100,9 +103,14 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
                     first = false;
                     try {
                         init(message.getUserId(), responseObserver);
-                        logger.info("Connected to user " + message.getUserId());
+
+                        observers.put(userId, responseObserver);
+                        groups.put(userId, groupId);
+
+                        logger.info("Initialized user " + message.getUserId());
                     } catch (SQLException | IOException e) {
-                        logger.warning("Couldn't init user connection => user " + message.getUserId() + " init rejected.");
+                        logger.warning("Couldn't init connection to user " + message.getUserId() + ".");
+                        removeUser(message.getUserId());
                         responseObserver.onError(new StatusRuntimeException(Status.ABORTED.withDescription("Couldn't init user.")));
                     }
                     return;
@@ -113,6 +121,7 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
                     newMessageId = saveMessage(message);
                 } catch (SQLException e) {
                     logger.warning("Couldn't save new message from user " + message.getUserId() + ".");
+                    removeUser(message.getUserId());
                     responseObserver.onError(new StatusRuntimeException(Status.ABORTED.withDescription("Couldn't save new message.")));
                     return;
                 }
@@ -136,38 +145,55 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
 
                 var response = responseBuilder.build();
 
-                for (var entry : observers.entrySet()) {
-                    if (entry.getKey() != message.getUserId()) {
+                Map<Long, StreamObserver<Message>> observersCopy;
+                Map<Long, Long> groupsCopy;
+                synchronized (this) {
+                    observersCopy = new HashMap<>(observers);
+                    groupsCopy = new HashMap<>(groups);
+                }
+
+                var toRemove = new LinkedList<Long>();
+
+                for (var entry : observersCopy.entrySet()) {
+                    if (groupsCopy.get(entry.getKey()) == groupId && entry.getKey() != message.getUserId()) {
                         try {
                             entry.getValue().onNext(response);
-                        } catch (StatusRuntimeException e) {
-                            logger.warning("Couldn't deliver message to user " + entry.getKey() + " => removing user.");
-                            entry.getValue().onError(new StatusRuntimeException(Status.ABORTED.withDescription("Couldn't deliver new message.")));
-                            removeUser(entry.getKey());
+                        } catch (StatusRuntimeException | IllegalStateException e) {
+                            logger.warning("Couldn't deliver message to user " + entry.getKey() + ".");
+                            toRemove.add(entry.getKey());
                         }
                     }
                 }
+
+                toRemove.forEach(this::removeUser);
 
                 logger.info("Received message " + newMessageId + " from user " + userId + " and sent to other users.");
             }
 
             private void init(long userId, StreamObserver<Message> responseObserver) throws SQLException, IOException {
-                var statement = connection.createStatement();
-                var resultSet = statement.executeQuery("SELECT group_id, ack_id FROM users WHERE user_id == " + userId);
+                ResultSet resultSet;
+
+                synchronized (this) {
+                    var statement = connection.createStatement();
+                    resultSet = statement.executeQuery("SELECT group_id, ack_id FROM users WHERE user_id == " + userId);
+                }
 
                 this.userId = userId;
                 this.groupId = resultSet.getLong(1);
                 var ackId = resultSet.getLong(2);
 
-                sendMissingMessages(this.groupId, ackId, responseObserver);
+                sendMissingMessages(ackId, responseObserver);
             }
 
-            private void sendMissingMessages(long groupId, long ackId, StreamObserver<Message> responseObserver) throws SQLException, IOException {
+            private synchronized void sendMissingMessages(long ackId, StreamObserver<Message> responseObserver) throws SQLException {
                 var statement = connection.createStatement();
                 var ackResultSet = statement.executeQuery("SELECT COALESCE(MAX(message_id), 0) FROM messages WHERE group_id == " + groupId);
                 var serverAckId = ackResultSet.getLong(1);
 
-                var resultsSet = statement.executeQuery("SELECT * FROM messages WHERE group_id == " + groupId + " AND message_id > " + ackId);
+                var resultsSet = statement.executeQuery(
+                        "SELECT * FROM messages WHERE group_id == " + groupId + " AND message_id > " + ackId + " AND user_id != " + userId + " ORDER BY time"
+                );
+
                 while (resultsSet.next()) {
                     var message = Message.newBuilder()
                             .setId(resultsSet.getLong("message_id"))
@@ -183,26 +209,23 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
                     }
 
                     var mime = resultsSet.getString("mime");
-                    var media = resultsSet.getBlob("media");
+                    var media = resultsSet.getBytes("media");
                     if (!resultsSet.wasNull()) {
                         message.setMime(mime);
-                        message.setMedia(ByteString.readFrom(media.getBinaryStream()));
+                        message.setMedia(ByteString.copyFrom(media));
                     }
 
                     responseObserver.onNext(message.build());
                 }
             }
 
-            private long saveMessage(Message message) throws SQLException {
+            private synchronized long saveMessage(Message message) throws SQLException {
                 var statement = connection.createStatement();
-                var idResult = statement.executeQuery("SELECT COALESCE(MAX(message_id), 0) FROM messages WHERE group_id == " + groupId);
-                var newMessageId = idResult.getLong(1) + 1;
-
                 statement.executeUpdate("UPDATE users SET ack_id = " + message.getAckId() + " WHERE user_id == " + userId);
                 removeOldMessages(groupId);
 
-                var preparedStatement = connection.prepareStatement("INSERT INTO messages VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)");
-                preparedStatement.setLong(1, newMessageId);
+                var preparedStatement = connection.prepareStatement("INSERT INTO messages VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)", Statement.RETURN_GENERATED_KEYS);
+                preparedStatement.setNull(1, Types.NULL);
                 preparedStatement.setLong(3, userId);
                 preparedStatement.setLong(4, groupId);
                 preparedStatement.setInt(5, message.getPriorityValue());
@@ -223,11 +246,17 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
                     preparedStatement.setNull(9, Types.NULL);
                 }
 
-                preparedStatement.executeUpdate();
-                return newMessageId;
+                var rowsNum = preparedStatement.executeUpdate();
+                var generatedKeys = statement.getGeneratedKeys();
+
+                if (rowsNum == 0 || !generatedKeys.next()) {
+                    throw new SQLException();
+                }
+
+                return generatedKeys.getLong(1);
             }
 
-            private void removeOldMessages(long groupId) throws SQLException {
+            private synchronized void removeOldMessages(long groupId) throws SQLException {
                 var statement = connection.createStatement();
                 var resultSet = statement.executeQuery("SELECT MIN(ack_id) FROM users WHERE group_id == " + groupId);
                 var ackId = resultSet.getLong(1);
@@ -241,7 +270,6 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
 
             @Override
             public void onCompleted() {
-                observers.get(userId).onCompleted();
                 removeUser(userId);
 
                 try {
@@ -249,18 +277,25 @@ public class ChatImpl extends ChatGrpc.ChatImplBase {
                 } catch (SQLException e) {
                     logger.warning("Couldn't remove old messages from group " + groupId + ".");
                 }
+
+                responseObserver.onCompleted();
+                logger.info("User " + userId + " disconnected.");
             }
 
-            private void removeUser(long userId) {
+            private synchronized void removeUser(long userId) {
+                if (observers.containsKey(userId)) {
+                    return;
+                }
+
+                observers.remove(userId);
+                groups.remove(userId);
+
                 try {
                     var statement = connection.createStatement();
                     statement.executeUpdate("DELETE FROM users WHERE user_id == " + userId);
                 } catch (SQLException e) {
                     logger.warning("Couldn't remove user " + userId + " from database.");
                 }
-
-                observers.remove(userId);
-                logger.info("User " + userId + " disconnected.");
             }
         };
     }
